Fix bugs TS, change name defaultValue to defaultSearchValue, useDebounce move to a separate file
