+++ src/hooks/useAction.ts
@@ -25,21 +25,3 @@ export const useAction = (action: () => PayloadAction<undefined>) => {
 
   return handler;
 };
-
-export const useDebounce = (value: string, delay: number) => {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-};
-
-
